backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
qdrant-client==1.8.0
google-generativeai==0.3.0
pyjwt==2.8.0
pytest==7.4.3
httpx==0.25.2
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0